+++ identity/auth/requests/users.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "@tanstack/vue-query";
 import { client } from "./client";
 import type { IdentityResponseBody } from "@saflib/identity-spec";
-import { handleClientMethod } from "@saflib/vue/tanstack";
-import { TanstackError } from "@saflib/vue/tanstack";
+import { handleClientMethod } from "@saflib/sdk";
+import { TanstackError } from "@saflib/sdk";
 export const useUsersQuery = () => {
   return useQuery<IdentityResponseBody["listUsers"][200], TanstackError>({
     queryKey: ["auth", "users"],
