+++ identity/auth/typed-mock.ts
@@ -1,4 +1,4 @@
-import { typedCreateHandler } from "@saflib/vue/testing";
+import { typedCreateHandler } from "@saflib/sdk/testing";
 import type { paths } from "@saflib/identity-spec";
 
 export const { createHandler: createIdentityHandler } =
