+++ sdk/docs/ref/@saflib/sdk/interfaces/ClientResult.md
@@ -1,4 +1,4 @@
-[**@saflib/vue**](../../../../index.md)
+[**@saflib/sdk**](../../../index.md)
 
 ***
