+++ sdk/docs/ref/testing/functions/typedCreateHandler.md
@@ -1,4 +1,4 @@
-[**@saflib/sdk**](../index.md)
+[**@saflib/sdk**](../../index.md)
 
 ***
