+++ sdk/docs/ref/testing/index.md
@@ -1,0 +1,11 @@
+[**@saflib/sdk**](../index.md)
+
+***
+
+# testing
+
+## Functions
+
+| Function | Description |
+| ------ | ------ |
+| [typedCreateHandler](functions/typedCreateHandler.md) | Use to create a typed helper function for creating typesafe mock API handlers. |
