+++ sdk/index.ts
@@ -1,11 +1,11 @@
 /**
  * Tanstack Query utilities for Vue.
- * @module @saflib/vue/tanstack
+ * @module @saflib/sdk
  */
 
 import { QueryClient } from "@tanstack/vue-query";
 import createClient from "openapi-fetch";
-import { isTestEnv } from "./src/env.ts";
+import { isTestEnv } from "./env.ts";
 
 /**
  * Given a "paths" openapi generated type and a subdomain, creates a typed `openapi-fetch` client which queries the given subdomain. Uses the current domain and protocol. Handles CSRF token injection, and works in tests.
