+++ sdk/testing/index.ts
@@ -1,4 +1,5 @@
 import { http, HttpResponse } from "msw";
+
 type ExtractRequestParams<Op extends Record<string, any>> =
   Op["parameters"] extends {
     path: { [key: string]: any };
