--- vue/docs/ref/@saflib/vue/functions/isTestEnv.md
@@ -1,11 +1,0 @@
-[**@saflib/vue**](../../../index.md)
-
-***
-
-# Function: isTestEnv()
-
-> **isTestEnv**(): `boolean`
-
-## Returns
-
-`boolean`
