+++ vue/docs/ref/@saflib/vue/type-aliases/LoaderQuery.md
@@ -4,7 +4,7 @@
 
 # Type Alias: LoaderQuery
 
-> **LoaderQuery** = `Pick`\<`UseQueryReturnType`\<`any`, [`TanstackError`](../tanstack/classes/TanstackError.md)\>, `"isLoading"` \| `"error"`\> & `object`
+> **LoaderQuery** = `Pick`\<`UseQueryReturnType`\<`any`, `TanstackError`\>, `"isLoading"` \| `"error"`\> & `object`
 
 A subset of what `useQuery` returns. This is so that loaders can create pseudo-queries by simply creating objects with isLoading, error, and isError properties.
