+++ vue/docs/ref/index.md
@@ -9,5 +9,4 @@
 | Module | Description |
 | ------ | ------ |
 | [@saflib/vue](@saflib/vue/index.md) | Common utilities for Vue SPAs and pages. |
-| [@saflib/vue/tanstack](@saflib/vue/tanstack/index.md) | Tanstack Query utilities for Vue. |
 | [@saflib/vue/testing](@saflib/vue/testing/index.md) | Testing utilities for Vue. |
