+++ vue/src/index.ts
@@ -5,7 +5,6 @@
 
 import "./assets.d.ts";
 export * from "./app.ts";
-export * from "./env.ts";
 export * from "./events.ts";
 export * from "./strings.ts";
 export * from "../types.ts";
