+++ vue/typedoc.json
@@ -1,4 +1,4 @@
 {
-  "entryPoints": ["./src/index.ts", "./tanstack.ts", "./testing/index.ts"],
+  "entryPoints": ["./src/index.ts", "./testing/index.ts"],
   "tsconfig": "./tsconfig.monorepo.json"
 }
